Add stellar sdk to project
The stellar-sdk package relies on node, e.g. the crypto package, so we can’t use this in a React Native environment. After trying out serveral solutions/hacks to make it work anyways, I just sided with a manual download of the browser package of the sdk.

It is taken from https://github.com/stellar/bower-js-stellar-sdk/blob/master/stellar-sdk.js. To match the import mechanism of React Native one has to manually add the following line to the end of the file:

export default StellarSdk;